move geographic focus to common
