+++ src/main/resources/templates/sparql/common/geographicFocus.sparql
