--- src/main/resources/templates/sparql/person/geographicFocus.sparql
@@ -1,10 +1,0 @@
-PREFIX vivo: <http://vivoweb.org/ontology/core#>
-PREFIX rdfs: <http://www.w3.org/2000/01/rdf-schema#>
-
-CONSTRUCT {
-    ?geographicFocus rdfs:label ?labelWithId .
-} WHERE {
-    <{{uri}}> vivo:geographicFocus ?geographicFocus .
-    ?geographicFocus rdfs:label ?label .
-    BIND( CONCAT( STR(?label), REPLACE(STR(?geographicFocus), "(^.*/)", "::") ) AS ?labelWithId ) .
-}
