--- src/main/resources/templates/sparql/process/geographicFocusType.sparql
@@ -1,11 +1,0 @@
-PREFIX vivo: <http://vivoweb.org/ontology/core#>
-PREFIX vitro: <http://vitro.mannlib.cornell.edu/ns/vitro/0.7#>
-
-CONSTRUCT {
-    ?geographicFocus vitro:mostSpecificType ?mostSpecificTypeWithId .
-} WHERE {
-    <{{uri}}> vivo:geographicFocus ?geographicFocus .
-    ?geographicFocus vitro:mostSpecificType ?mostSpecificType .
-    BIND( CONCAT( STR(?mostSpecificType), REPLACE(STR(?geographicFocus), "(^.*/)", "::") ) AS ?mostSpecificTypeWithId ) .
-    FILTER (STRSTARTS (STR(?mostSpecificType), STR(vivo:))) .
-}
