--- src/main/resources/templates/sparql/relationship/geographicFocus.sparql
@@ -1,10 +1,0 @@
-PREFIX rdfs: <http://www.w3.org/2000/01/rdf-schema#>
-PREFIX vivo: <http://vivoweb.org/ontology/core#>
-
-CONSTRUCT {
-    ?geographicFocus rdfs:label ?labelWithId .
-} WHERE {
-    <{{uri}}> vivo:geographicFocus ?geographicFocus ; a vivo:Grant .
-    ?geographicFocus rdfs:label ?label ; a vivo:GeographicRegion .
-    BIND( CONCAT( STR(?label), REPLACE(STR(?geographicFocus), "(^.*/)", "::") ) AS ?labelWithId ) .
-}
