--- src/main/resources/templates/sparql/relationship/geographicFocusType.sparql
@@ -1,11 +1,0 @@
-PREFIX skos: <http://www.w3.org/2004/02/skos/core#>
-PREFIX vitro: <http://vitro.mannlib.cornell.edu/ns/vitro/0.7#>
-PREFIX vivo: <http://vivoweb.org/ontology/core#>
-
-CONSTRUCT {
-    ?geographicFocus vitro:mostSpecificType ?mostSpecificTypeWithId .
-} WHERE {
-    <{{uri}}> vivo:geographicFocus ?geographicFocus ; a vivo:Grant .
-    ?geographicFocus vitro:mostSpecificType ?mostSpecificType ; a vivo:GeographicRegion .
-    BIND( CONCAT( STR(?mostSpecificType), REPLACE(STR(?geographicFocus), "(^.*/)", "::") ) AS ?mostSpecificTypeWithId ) .
-}
